+++ checkpoint/orbax/checkpoint/_src/metadata/checkpoint.py
@@ -50,7 +50,7 @@ def _sanitize_metadata_path(path: epath.PathLike) -> epath.Path:
 
 
 def step_metadata_file_path(path: epath.PathLike) -> epath.Path:
-  """The path to step metadata file for a given checkpoint directory."""
+  """The path to step metadata file, `_CHECKPOINT_METADATA`."""
   return _sanitize_metadata_path(path) / _STEP_METADATA_FILENAME