+++ checkpoint/orbax/checkpoint/_src/path/format_utils.py
@@ -34,6 +34,7 @@ def _has_msgpack_metadata_file(path: epath.Path) -> bool:
 
 
 def _has_pytree_metadata_file(path: epath.Path) -> bool:
+  """Returns True if path contains `_METADATA` or `checkpoint` file."""
   return (path / PYTREE_METADATA_FILE).exists() or _has_msgpack_metadata_file(
       path
   )