+++ checkpoint/orbax/checkpoint/experimental/v1/_src/path/format_utils.py
@@ -53,21 +53,12 @@ def validate_pytree_checkpoint(path: path_types.PathLike):
     raise FileNotFoundError(f'Checkpoint path {path} does not exist.')
   if not path.is_dir():
     raise NotADirectoryError(f'Checkpoint path {path} is not a directory.')
-  metadata_store = checkpoint_metadata.metadata_store(enable_write=False)
-  # Path points to a single step checkpoint with valid metadata.
-  checkpoint_metadata_path = checkpoint_metadata.step_metadata_file_path(path)
-  if not checkpoint_metadata_path.exists():
-    raise FileNotFoundError(
-        f'Checkpoint path {path} does not contain a valid metadata file.'
-    )
-  if metadata_store.read(checkpoint_metadata_path) is None:
-    raise ValueError(
-        f'Failed to read valid metadata for checkpoint path {path}.'
-    )
   if not (path / PYTREE_CHECKPOINTABLE_KEY).exists():
     raise FileNotFoundError(
-        f'Checkpoint path {path} does not contain a PyTree checkpointable'
-        f' (called "{PYTREE_CHECKPOINTABLE_KEY}").'
+        f'Checkpoint path {path} must contain a subdirectory named'
+        f' "{PYTREE_CHECKPOINTABLE_KEY}". Please try inspecting the'
+        ' checkpointable metadata using `ocp.checkpointables_metadata()` or'
+        ' try loading the checkpoint using `ocp.load_checkpointables()`.'
     )
   if not format_utils._has_pytree_metadata_file(  # pylint: disable=protected-access
       path / PYTREE_CHECKPOINTABLE_KEY
@@ -83,4 +74,16 @@ def validate_pytree_checkpoint(path: path_types.PathLike):
         ' sign of a malformed checkpoint, unless your checkpoint consists'
         ' entirely of strings or other non-standard PyTree leaves.',
         path,
+    )
+  metadata_store = checkpoint_metadata.metadata_store(enable_write=False)
+  # Path points to a single step checkpoint with valid metadata.
+  checkpoint_metadata_path = checkpoint_metadata.step_metadata_file_path(path)
+  if not checkpoint_metadata_path.exists():
+    raise FileNotFoundError(
+        f'Checkpoint path {path} does not contain a valid metadata file:'
+        f' {checkpoint_metadata_path.name}'
+    )
+  if metadata_store.read(checkpoint_metadata_path) is None:
+    raise ValueError(
+        f'Failed to read valid metadata for checkpoint path {path}.'
     )