+++ checkpoint/CHANGELOG.md
@@ -22,10 +22,8 @@ would want them to be preferred.
 - #v1 Add `JsonHandler`.
 - #v1 Add `training.Checkpointer`.
 - #v1 Add checkpointables support for `training.Checkpointer`.
-
-### Added
-
 - `PartsOf` structure which holds a PyTree whose leaf nodes may be missing.
+- #v1 Add compatibility tests for save-by-v0-load-by-v1 and also fix code.
 
 ## [0.11.12] - 2025-04-09